[repo:adventofcode] go/2015/14: add placeholder and benchmark for part 1
There are no tests since the example given in the problem description
does not have an answer for the specific number of seconds we are
supposed to use (2503 seconds), but only for 1000 seconds. One could
probably make the number of seconds a function argument somewhere, but
I view it as an invariant of the problem for now.