[repo:adventofcode] go/2015/19: add functions to parse input into richer data structures
I unfortunately have to make one assumption based on my input (and
which is not described in the assignment): a substance is an upper
rune followed by zero or more lower runes. However, there is one
replacement specified in the input where the replaced substance is a
single lower rune: `e => ...`. However, that substance is not part of
any ohter replacements, or part of the molecule to replace, so I think
my heuristic will work.