[repo:adventofcode] cpp/2019/12: optimize solution a bit
We will always eventually reach the initial state, and therefore we do
not need to keep track of all states we have visited. This is what I
had in mind all the time (since that is why I do the modulo logic when
finding the state after `n` steps), but for some reason overlooked it
in this specific part of the program. This reduces the runtime by
about 10x, from 650 ms to about 70 ms (both numbers are with compiler
optimizations enabled).