tasks: add a bunch of test cases in a Markdown file
The purpose is only to have written down my ideas for various test
cases somewhere before I have actually implemented them as tests, so
that I don't lose track of the ideas.

Also, this list is not complete: there are a bunch of methods missing,
and I will surely come up with new test cases for the methods that are
listed.