tasks/server: remove current implementation and tests
With the new list of test cases (see tasks/testcases.md), I figured I
might as well throw out the existing tests -- they were only
proof-of-concept, after all. The only test that remains is a test for
the `setup` function, which will act as a starting point for writing
the new tests.

Also update external dependencies as appropriate.