testing/grpctest: fix flake in NewClientConnT
The bug was that the call to grpc.Dial did not include the
grpc.WithBlock option. This sometimes caused tests that call
NewClientConnT to finish before the gRPC server had even started
up, meaning that s.GracefulStop was called before s.Serve had started
up the server. This meant that s.Serve returned grpc.ErrServerStopped,
and the goroutine running the gRPC server would call t.Errorf after
the test had finished -- which results in a panic.

While adding grpc.WithBlock to the call to grpc.Dial should be enough
to fix this bug, I also took the opportunity to clean up the test some
more -- notably cleaning up the *grpc.ClientConn, and running the gRPC
server in an errgroup.Group (which is a great package and type).