aip/fieldmask: add package fieldmask with a simple Merge function
The purpose of the Merge function is to update a proto message based
on a field mask. Essentially, it works like the ordinary proto.Merge
function, except it can do partial updates.

Initially, it only supports scalar fields. Support for nested
messages, repeated fields, and maps, will be added as needed.