Remove Bazel build files and generate Go code manually
I'm really beginning to feel that using Bazel is turning out to be
more trouble than it's worth, at least for now. Just using manual code
generation and the `go` tools is definitely good enough for now. As a
bonus, now `gopls` will be able to work with the protos.

In the future I might bring Bazel back, e.g., when it's time to begin
to build multi-language stuff (say, a JS frontend).