tasks2: define an alternative `Tasks` service
This is a second attempt at defining a tasks service. It includes many
more RPCs, an event resource to track updates to a task, as well as a
"deferral" concept that tries to model that tasks can be deferred
until some conditions are fulfilled, such as all dependency tasks
being completed or a certain timestamp has been passed.