adventofcode/go: make all solutions build and pass tests
This involved quite a bit of code editing, mainly involving where to
find input files. Also, the internal/geo package had been broken for
quite some time (it failed to build the tests) and the 2015/03
solution used it, so it had to be changed as well.

I considered not using the `bazel` package and instead setting `rundir
= "."` in the `go_test` rules, to make Go tests behave like the Java
and C++ tests. However, that would be less maintainable in the long
term since Gazelle doesn't set it by default (and I don't think
there's a way to configure it to do that either).