WORKSPACE: update to an unreleased version of rules_go
This new version includes bazelbuild/rules_go#3029 which has the effect that
building a benchmark with `-c dbg` and then running with
`-test.cpuprofile=<whatever>` lets `go tool pprof` see source information. This
is not the case with the most recently released version of rules_go, v0.29.0.

Also update Makefile so that `make buildifier` formats WORKSPACE.